+++ CHANGELOG.md
@@ -15,6 +15,7 @@ To see unreleased changes, please see the [CHANGELOG on the master branch](https
 ### Added
 
 * config: `project` option.
+* config: `service.<name>.scale` option.
 * jinja2 templates for configs.
 * Perform all migrations and collections loading.